feat: enrich disantrefact MusicAlbum json-ld with streaming, genre and artist links
