+++ src/pages/DisantrefactPage/data.js
@@ -1,43 +1,5 @@
 export const YT_VIDEO_ID = 'Q2ihmC8a_RE';
 
-export const STREAM_LINKS = [
-  {
-    label: 'Spotify',
-    icon: 'spotify',
-    href: 'https://open.spotify.com/album/48INkrzUDxeVtKj5fqn3Ub?si=_bpUuBZTTP-LIxEGLnCV2g',
-  },
-  {
-    label: 'Apple Music',
-    icon: 'appleMusic',
-    href: 'https://music.apple.com/au/album/disantrefact/6768293725',
-  },
-  {
-    label: 'YouTube Music',
-    icon: 'youtubeMusic',
-    href: 'https://music.youtube.com/playlist?list=OLAK5uy_lUHQkmJlijq2r-47lTh53s_Dg7uRW1BKA&si=VjiSa0-l41HHkjGa',
-  },
-  {
-    label: 'Deezer',
-    icon: 'deezer',
-    href: 'https://link.deezer.com/s/33wa9nzAdSzJeVDSkWm41',
-  },
-  {
-    label: 'Tidal',
-    icon: 'tidal',
-    href: 'https://tidal.com/album/523459824/u',
-  },
-  {
-    label: 'Bandcamp',
-    icon: 'bandcamp',
-    href: 'https://cutmylips.bandcamp.com/album/disantrefact',
-  },
-  {
-    label: 'YouTube',
-    icon: 'youtube',
-    href: 'https://youtu.be/Tm_GbJLUew4',
-  },
-];
-
 export const SOCIAL_LINKS = [
   {
     label: 'Instagram',
